requirements.txt
# HTTP client (snowflake ID service)
requests>=2.28.0

# Django and related packages
Django==4.1.13
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.11.0

# Database
pymysql>=1.1.0

# ML and scientific
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
numpy>=1.24.0

# Config and optimization
hydra-core>=1.3.0
PyYAML>=6.0
optuna>=3.4.0
mlflow>=2.9.0

# Optional: Ray for distributed execution
# ray>=2.9.0

# AI (OpenAI-compatible API)
openai>=1.0.0
